Revisit test, provide improvements
