+++ core/src/main/resources/alert-manager-tpt.json
